Начал переход на чистую архитектуру, раскидал всё пакетам data, domain, presentation. Написал domain слой.
