--- app/src/main/java/com/example/cryptoapp/adapter/holder/CoinInfoViewHolder.kt
@@ -1,8 +1,0 @@
-package com.example.cryptoapp.adapter.holder
-
-import androidx.recyclerview.widget.RecyclerView.ViewHolder
-import com.example.cryptoapp.databinding.ItemCoinInfoBinding
-
-class CoinInfoViewHolder(
-    val binding: ItemCoinInfoBinding
-) : ViewHolder(binding.root)
