--- app/src/main/java/com/example/cryptoapp/api/ApiFactory.kt
@@ -1,19 +1,0 @@
-package com.example.cryptoapp.api
-
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-
-object ApiFactory {
-
-    private const val BASE_URl = "https://min-api.cryptocompare.com/data/"
-    const val BASE_IMAGE_URl = "https://cryptocompare.com"
-
-    private val retrofit = Retrofit.Builder()
-        .addConverterFactory(GsonConverterFactory.create())
-        .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
-        .baseUrl(BASE_URl)
-        .build()
-
-    val apiService = retrofit.create(ApiService::class.java)
-}
