+++ app/src/main/java/com/example/cryptoapp/data/database/AppDatabase.kt
@@ -1,0 +1,32 @@
+package com.example.cryptoapp.data.database
+
+import android.app.Application
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import com.example.cryptoapp.data.model.priceinfo.CoinPriceInfo
+
+@Database(entities = [CoinPriceInfo::class], version = 1, exportSchema = false)
+abstract class AppDatabase : RoomDatabase() {
+
+    abstract fun coinPriceInfoDao(): CoinPriceInfoDao
+
+    companion object {
+        private const val NAME_DB = "main.db"
+        private var instance: AppDatabase? = null
+        private val LOCK = Any()
+
+        fun getInstance(application: Application): AppDatabase {
+            synchronized(LOCK) {
+                instance?.let { return it }
+                val db = Room.databaseBuilder(
+                    application,
+                    AppDatabase::class.java,
+                    NAME_DB
+                ).build()
+                instance = db
+                return db
+            }
+        }
+    }
+}
