+++ app/src/main/java/com/example/cryptoapp/data/database/CoinPriceInfoDao.kt
@@ -1,0 +1,20 @@
+package com.example.cryptoapp.data.database
+
+import androidx.lifecycle.LiveData
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import com.example.cryptoapp.data.model.priceinfo.CoinPriceInfo
+
+@Dao
+interface CoinPriceInfoDao {
+    @Query("SELECT * FROM full_price_list ORDER BY lastUpdate desc")
+    fun getPriceList(): LiveData<List<CoinPriceInfo>>
+
+    @Query("SELECT * FROM full_price_list WHERE fromSymbol == :fSym LIMIT 1")
+    fun getPriceInfoAboutCoin(fSym: String): LiveData<CoinPriceInfo>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    fun insertPriceList(priceList: List<CoinPriceInfo>)
+}
