+++ app/src/main/java/com/example/cryptoapp/data/model/coininfo/CoinInfo.kt
@@ -1,0 +1,10 @@
+package com.example.cryptoapp.data.model.coininfo
+
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+data class CoinInfo(
+    @SerializedName("Name")
+    @Expose
+    val name: String? = null
+)
