+++ app/src/main/java/com/example/cryptoapp/data/model/coininfo/Datum.kt
@@ -1,0 +1,10 @@
+package com.example.cryptoapp.data.model.coininfo
+
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+data class Datum(
+    @SerializedName("CoinInfo")
+    @Expose
+    val coinInfo: CoinInfo? = null
+)
