+++ app/src/main/java/com/example/cryptoapp/data/model/news/NewsResponse.kt
@@ -1,0 +1,8 @@
+package com.example.cryptoapp.data.model.news
+
+import com.google.gson.annotations.SerializedName
+
+data class NewsResponse(
+    @SerializedName("Data")
+    val data: List<Data>?
+)
