+++ app/src/main/java/com/example/cryptoapp/data/model/news/SourceInfo.kt
@@ -1,0 +1,9 @@
+package com.example.cryptoapp.data.model.news
+
+import com.google.gson.annotations.SerializedName
+import java.io.Serializable
+
+data class SourceInfo(
+    @SerializedName("name")
+    val name: String?
+) : Serializable
