+++ app/src/main/java/com/example/cryptoapp/data/model/priceinfo/CoinPriceInfoRawData.kt
@@ -1,0 +1,11 @@
+package com.example.cryptoapp.data.model.priceinfo
+
+import com.google.gson.JsonObject
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+data class CoinPriceInfoRawData(
+    @SerializedName("RAW")
+    @Expose
+    val coinPriceJsonObject: JsonObject? = null
+)
