--- app/src/main/java/com/example/cryptoapp/database/CoinPriceInfoDao.kt
@@ -1,20 +1,0 @@
-package com.example.cryptoapp.database
-
-import androidx.lifecycle.LiveData
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.example.cryptoapp.pojo.priceinfo.CoinPriceInfo
-
-@Dao
-interface CoinPriceInfoDao {
-    @Query("SELECT * FROM full_price_list ORDER BY lastUpdate desc")
-    fun getPriceList(): LiveData<List<CoinPriceInfo>>
-
-    @Query("SELECT * FROM full_price_list WHERE fromSymbol == :fSym LIMIT 1")
-    fun getPriceInfoAboutCoin(fSym: String): LiveData<CoinPriceInfo>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insertPriceList(priceList: List<CoinPriceInfo>)
-}
