+++ app/src/main/java/com/example/cryptoapp/domain/entity/CoinInfo.kt
@@ -1,0 +1,12 @@
+package com.example.cryptoapp.domain.entity
+
+data class CoinInfo(
+    val fromSymbol: String,
+    val toSymbol: String?,
+    val price: String?,
+    val lastUpdate: Long?,
+    val highDay: String?,
+    val lowDay: String?,
+    val lastMarket: String?,
+    val imageUrl: String?
+)
