+++ app/src/main/java/com/example/cryptoapp/domain/usecase/GetCoinInfoUseCase.kt
@@ -1,0 +1,10 @@
+package com.example.cryptoapp.domain.usecase
+
+import com.example.cryptoapp.domain.repository.CoinRepository
+
+class GetCoinInfoUseCase(
+    private val repository: CoinRepository
+) {
+
+    operator fun invoke(fromSymbol: String) = repository.getCoinInfo(fromSymbol)
+}
