--- app/src/main/java/com/example/cryptoapp/pojo/coininfo/CoinInfo.kt
@@ -1,10 +1,0 @@
-package com.example.cryptoapp.pojo.coininfo
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-data class CoinInfo(
-    @SerializedName("Name")
-    @Expose
-    val name: String? = null
-)
