--- app/src/main/java/com/example/cryptoapp/pojo/coininfo/CoinInfoListOfData.kt
@@ -1,10 +1,0 @@
-package com.example.cryptoapp.pojo.coininfo
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-data class CoinInfoListOfData(
-    @SerializedName("Data")
-    @Expose
-    val data: List<Datum>? = null
-)
