--- app/src/main/java/com/example/cryptoapp/pojo/coininfo/Datum.kt
@@ -1,10 +1,0 @@
-package com.example.cryptoapp.pojo.coininfo
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-data class Datum(
-    @SerializedName("CoinInfo")
-    @Expose
-    val coinInfo: CoinInfo? = null
-)
