--- app/src/main/java/com/example/cryptoapp/pojo/news/Data.kt
@@ -1,19 +1,0 @@
-package com.example.cryptoapp.pojo.news
-
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-data class Data(
-    @SerializedName("id")
-    val id: Int,
-    @SerializedName("guid")
-    val guid: String?,
-    @SerializedName("imageurl")
-    val imageUrl: String?,
-    @SerializedName("title")
-    val title: String?,
-    @SerializedName("body")
-    val body: String?,
-    @SerializedName("source_info")
-    val sourceInfo: SourceInfo?
-) : Serializable
