--- app/src/main/java/com/example/cryptoapp/pojo/news/NewsResponse.kt
@@ -1,8 +1,0 @@
-package com.example.cryptoapp.pojo.news
-
-import com.google.gson.annotations.SerializedName
-
-data class NewsResponse(
-    @SerializedName("Data")
-    val data: List<Data>?
-)
