--- app/src/main/java/com/example/cryptoapp/pojo/news/SourceInfo.kt
@@ -1,9 +1,0 @@
-package com.example.cryptoapp.pojo.news
-
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-data class SourceInfo(
-    @SerializedName("name")
-    val name: String?
-) : Serializable
