--- app/src/main/java/com/example/cryptoapp/pojo/priceinfo/CoinPriceInfoRawData.kt
@@ -1,11 +1,0 @@
-package com.example.cryptoapp.pojo.priceinfo
-
-import com.google.gson.JsonObject
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-data class CoinPriceInfoRawData(
-    @SerializedName("RAW")
-    @Expose
-    val coinPriceJsonObject: JsonObject? = null
-)
