+++ app/src/main/java/com/example/cryptoapp/presentation/adapter/holder/CoinInfoViewHolder.kt
@@ -1,0 +1,8 @@
+package com.example.cryptoapp.presentation.adapter.holder
+
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import com.example.cryptoapp.databinding.ItemCoinInfoBinding
+
+class CoinInfoViewHolder(
+    val binding: ItemCoinInfoBinding
+) : ViewHolder(binding.root)
