+++ app/src/main/java/com/example/cryptoapp/presentation/adapter/holder/NewsViewHolder.kt
@@ -1,0 +1,8 @@
+package com.example.cryptoapp.presentation.adapter.holder
+
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import com.example.cryptoapp.databinding.ItemNewsBinding
+
+class NewsViewHolder(
+    val binding: ItemNewsBinding
+) : ViewHolder(binding.root)
