+++ app/src/main/res/layout/activity_coin_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/light_grey"
-    tools:context=".view.CoinDetailActivity">
+    tools:context=".presentation.view.CoinDetailActivity">
 
     <ImageView
         android:id="@+id/ivCoinLogo"
