+++ app/src/main/res/layout/activity_coin_price_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CoinPriceListActivity"
+    tools:context=".presentation.view.CoinPriceListActivity"
     android:background="@color/light_grey">
 
     <androidx.recyclerview.widget.RecyclerView
