+++ app/src/main/res/layout/activity_detail_news.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/light_grey"
-    tools:context=".view.DetailNewsActivity">
+    tools:context=".presentation.view.DetailNewsActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
