+++ app/src/main/res/layout/activity_news_coin.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/light_grey"
-    tools:context=".view.NewsCoinActivity">
+    tools:context=".presentation.view.NewsCoinActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewNews"
